requirements.txt
lxml==4.4.2
matplotlib==3.1.1
numpy==1.17.4
pandas==0.25.3
requests==2.22.0
opencv_python==4.1.2.30
Pillow==6.2.1
